Move FileFinderView into its own module
