+++ isobar_core/src/lib.rs
@@ -16,6 +16,7 @@ pub mod notify_cell;
 pub mod window;
 pub mod workspace;
 pub mod fuzzy_search;
+mod file_finder;
 
 mod movement;
 mod tree;
